Seed user's profiles and display user's names correctly in all views

+++ database/migrations/2019_09_21_012454_create_profiles_table.php
@@ -15,7 +15,7 @@ class CreateProfilesTable extends Migration
     {
         Schema::create('profiles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('user_id');
+            $table->unsignedInteger('user_id');
             $table->string('first_name');
             $table->string('middle_name');
             $table->string('last_name');
